Modified ApiTests and Added Structure for Account Types
The ApiTests page can now accept test from any component that wishes to use it. You need to make a component that has api calls or a component that depends on an api.js file.

The layouts are structured so that Kris can fill in the rest of the work.