Get/set user's role upon logging in, navbar changes, and loading modal.
Because of this, the appropriate navigation bar can now be displayed and has been coded to do so. Additionally, a log out button was finally added to successfully terminate the current session. Lastly, a loading modal was added to add confirmation that something is working as opposed to nothing happening on screen.